requirements.txt
streamlit
pandas
playwright
wikipedia-api
requests
beautifulsoup4
googlesearch-python
google-generativeai
python-dotenv
openpyxl
